refactor: 불필요한 Service 인터페이스 삭제
