+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -1,14 +1,48 @@
 package roomescape.reservation.service;
 
 import java.util.List;
+import org.springframework.stereotype.Service;
 import roomescape.reservation.dto.ReservationRequest;
 import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.entity.Reservation;
+import roomescape.reservation.entity.ReservationTime;
+import roomescape.common.exception.EntityNotFoundException;
+import roomescape.reservation.repository.ReservationRepository;
+import roomescape.reservation.repository.ReservationTimeRepository;
 
-public interface ReservationService {
+@Service
+public class ReservationService {
 
-    List<ReservationResponse> getAll();
+    private final ReservationRepository reservationRepository;
+    private final ReservationTimeRepository reservationTimeRepository;
 
-    ReservationResponse save(ReservationRequest request);
+    public ReservationService(ReservationRepository reservationRepository,
+                              ReservationTimeRepository reservationTimeRepository) {
+        this.reservationRepository = reservationRepository;
+        this.reservationTimeRepository = reservationTimeRepository;
+    }
 
-    void delete(Long id);
+    public List<ReservationResponse> getAll() {
+        List<Reservation> reservations = reservationRepository.findAll();
+
+        return reservations.stream()
+                .map(ReservationResponse::from)
+                .toList();
+    }
+
+    public ReservationResponse save(ReservationRequest request) {
+        Long timeId = request.timeId();
+        ReservationTime reservationTime = reservationTimeRepository.findById(timeId)
+                .orElseThrow(() -> new EntityNotFoundException("reservationsTime not found id =" + timeId));
+
+        Reservation reservation = Reservation.withoutId(request.name(), request.date(), reservationTime);
+
+        Reservation saved = reservationRepository.save(reservation);
+
+        return ReservationResponse.from(saved);
+    }
+
+    public void delete(Long id){
+        reservationRepository.deleteById(id);
+    }
 }
