+++ src/main/java/roomescape/reservation/service/ReservationTimeService.java
@@ -1,14 +1,38 @@
 package roomescape.reservation.service;
 
 import java.util.List;
+import org.springframework.stereotype.Service;
 import roomescape.reservation.dto.ReservationTimeRequest;
 import roomescape.reservation.dto.ReservationTimeResponse;
+import roomescape.reservation.entity.ReservationTime;
+import roomescape.reservation.repository.ReservationTimeRepository;
 
-public interface ReservationTimeService {
+@Service
+public class ReservationTimeService {
 
-    List<ReservationTimeResponse> getAll();
+    private final ReservationTimeRepository reservationTimeRepository;
 
-    ReservationTimeResponse save(ReservationTimeRequest request);
+    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+        this.reservationTimeRepository = reservationTimeRepository;
+    }
 
-    void delete(Long id);
+    public List<ReservationTimeResponse> getAll() {
+        List<ReservationTime> reservationTimes = reservationTimeRepository.findAll();
+
+        return reservationTimes.stream()
+                .map(ReservationTimeResponse::from)
+                .toList();
+    }
+
+    public ReservationTimeResponse save(ReservationTimeRequest request) {
+        ReservationTime reservationTime = ReservationTime.withoutId(request.startAt());
+
+        ReservationTime saved = reservationTimeRepository.save(reservationTime);
+
+        return ReservationTimeResponse.from(saved);
+    }
+
+    public void delete(Long id){
+        reservationTimeRepository.deleteById(id);
+    }
 }
