--- src/main/java/roomescape/reservation/service/impl/ReservationTimeServiceImpl.java
@@ -1,42 +1,0 @@
-package roomescape.reservation.service.impl;
-
-import java.util.List;
-import org.springframework.stereotype.Service;
-import roomescape.reservation.dto.ReservationTimeRequest;
-import roomescape.reservation.dto.ReservationTimeResponse;
-import roomescape.reservation.entity.ReservationTime;
-import roomescape.reservation.repository.ReservationTimeRepository;
-import roomescape.reservation.service.ReservationTimeService;
-
-@Service
-public class ReservationTimeServiceImpl implements ReservationTimeService {
-
-    private final ReservationTimeRepository reservationTimeRepository;
-
-    public ReservationTimeServiceImpl(ReservationTimeRepository reservationTimeRepository) {
-        this.reservationTimeRepository = reservationTimeRepository;
-    }
-
-    @Override
-    public List<ReservationTimeResponse> getAll() {
-        List<ReservationTime> reservationTimes = reservationTimeRepository.findAll();
-
-        return reservationTimes.stream()
-                .map(ReservationTimeResponse::from)
-                .toList();
-    }
-
-    @Override
-    public ReservationTimeResponse save(ReservationTimeRequest request) {
-        ReservationTime reservationTime = ReservationTime.withoutId(request.startAt());
-
-        ReservationTime saved = reservationTimeRepository.save(reservationTime);
-
-        return ReservationTimeResponse.from(saved);
-    }
-
-    @Override
-    public void delete(Long id){
-        reservationTimeRepository.deleteById(id);
-    }
-}
